+++ includes/libraries/aihrus-framework/CHANGELOG.md
@@ -3,9 +3,20 @@
 ## master
 
 ## 1.1.5RC1
+* Abstract do_validate_settings for easier reuse
+* Add widget's form_parts method
 * Don't shortcode attribute for reset options
+* Move get_suggest to widget abstract class from interface
+* RESOLVE Add widget option defaults
+* RESOLVE form_parts not allowing no arguments
 * RESOLVE michael-cannon/testimonials-widget#6 On widget options, switch labels expand/collapse when using them
 * RESOLVE michael-cannon/testimonials-widget#78 Show revert link in deactivation admin notice
+* RESOLVE Move form_instance to abstract class Aihrus_Widget
+* RESOLVE Move get_suggest to abstract class Aihrus_Widget
+* RESOLVE Remove tw- reference
+* RESOLVE Space missing before Collapse
+* RESOLVE Use widget args than globals
+* Revise widget class for easier reuse
 * Update licensing GPL 2.0
 
 ## 1.1.4